+++ app/controllers/sign_in_controller.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'sign_in/logingov/service'
+require 'sign_in/idme/service'
 
 class SignInController < ApplicationController
   skip_before_action :verify_authenticity_token
