+++ config/settings/test.yml
@@ -21,16 +21,21 @@ reports:
 logingov:
   client_id: urn:gov:gsa:openidconnect.profiles:sp:sso:va:dev_signin
   redirect_uri: http://localhost:3000/sign_in/logingov/callback
+<<<<<<< HEAD
   client_cert_path: spec/fixtures/sign_in/logingov.crt
   client_key_path: spec/fixtures/sign_in/logingov.key
+=======
+  client_cert_path: ~
+  client_key_path: ~
+>>>>>>> Changes from huddle feedback
   oauth_url: https://idp.int.identitysandbox.gov
 
 idme:
   client_id: dde0b5b8bfc023a093830e64ef83f148
-  redirect_uri: http://localhost:3000/v2/sessions/idme/callback
+  redirect_uri: http://localhost:3000/sign_in/idme/callback
   client_cert_path: ~
   client_key_path: ~
-  oauth_url: https://api.idmelabs.com/oauth/authorize?client_id=dde0b5b8bfc023a093830e64ef83f148&redirect_uri=http://localhost:3000/v2/sessions/callback&response_type=code&scope=http://idmanagement.gov/ns/assurance/loa/3
+  oauth_url: https://api.idmelabs.com
 
 saml:
   cert_path: spec/support/certificates/ruby-saml.crt
